# ai-specs/specs/api-spec.yml
openapi: 3.1.0
info:
  title: Quermed CRM API
  version: 1.0.0
paths:
...
  /api/v1/opportunities:
    get:
      tags:
      - opportunities
      summary: List opportunities (scoped)
      operationId: list_opportunities_api_v1_opportunities_get
      security:
      - HTTPBearer: []
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: pipeline_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Pipeline Id
      - name: stage_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Stage Id
      - name: division_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Division Id
      - name: owner_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Owner Id
      - name: account_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Account Id
      - name: is_tender
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Tender
      - name: is_at_risk
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is At Risk
      - name: close_from
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Close From
      - name: close_to
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Close To
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
          title: Q
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Page Size
      - name: sort
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Sort
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_OpportunitySummaryRead_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - opportunities
      summary: Create an opportunity (three fields, smart defaults)
      operationId: create_opportunity_api_v1_opportunities_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpportunityCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpportunityRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/opportunities/{opportunity_id}:
    get:
      tags:
      - opportunities
      summary: Opportunity detail
      operationId: read_opportunity_api_v1_opportunities__opportunity_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: opportunity_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Opportunity Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpportunityRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - opportunities
      summary: Update descriptive fields
      operationId: update_opportunity_api_v1_opportunities__opportunity_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: opportunity_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Opportunity Id
      - name: If-Match
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: If-Match
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpportunityUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpportunityRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/quotes:
    get:
      tags:
      - quotes
      summary: List quotes (current versions)
      operationId: list_quotes_api_v1_quotes_get
      security:
      - HTTPBearer: []
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: owner_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Owner Id
      - name: opportunity_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Opportunity Id
      - name: account_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Account Id
      - name: expiring
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Expiring
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
          title: Q
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Page Size
      - name: sort
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Sort
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_QuoteSummaryRead_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - quotes
      summary: Create a draft from an opportunity (copies its lines)
      operationId: create_quote_api_v1_quotes_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuoteCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/QuoteRead'
                - $ref: '#/components/schemas/QuotePublicRead'
                title: Response Create Quote Api V1 Quotes Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/quotes/{quote_id}:
    get:
      tags:
      - quotes
      summary: Quote detail
      operationId: read_quote_api_v1_quotes__quote_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: quote_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Quote Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/QuoteRead'
                - $ref: '#/components/schemas/QuotePublicRead'
                title: Response Read Quote Api V1 Quotes  Quote Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - quotes
      summary: Update a draft
      operationId: update_quote_api_v1_quotes__quote_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: quote_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Quote Id
      - name: If-Match
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: If-Match
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuoteUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/QuoteRead'
                - $ref: '#/components/schemas/QuotePublicRead'
                title: Response Update Quote Api V1 Quotes  Quote Id  Patch
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - quotes
      summary: Delete a draft
      operationId: delete_quote_api_v1_quotes__quote_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: quote_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Quote Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/quotes/{quote_id}/pdf:
    get:
      tags:
      - quotes
      summary: Quote PDF (stored when sent, preview for drafts)
      operationId: read_quote_pdf_api_v1_quotes__quote_id__pdf_get
      security:
      - HTTPBearer: []
      parameters:
      - name: quote_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Quote Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/quotes/{quote_id}/send:
    post:
      tags:
      - quotes
      summary: Freeze the version, store the PDF and email it
      operationId: send_quote_api_v1_quotes__quote_id__send_post
      security:
      - HTTPBearer: []
      parameters:
      - name: quote_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Quote Id
      - name: If-Match
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: If-Match
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuoteSend'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/QuoteRead'
                - $ref: '#/components/schemas/QuotePublicRead'
                title: Response Send Quote Api V1 Quotes  Quote Id  Send Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/quotes/{quote_id}/accept:
    post:
      tags:
      - quotes
      summary: 'Accept: wins the opportunity and rejects sibling quotes'
      operationId: accept_quote_api_v1_quotes__quote_id__accept_post
      security:
      - HTTPBearer: []
      parameters:
      - name: quote_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Quote Id
      - name: If-Match
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: If-Match
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuoteAccept'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/QuoteRead'
                - $ref: '#/components/schemas/QuotePublicRead'
                title: Response Accept Quote Api V1 Quotes  Quote Id  Accept Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/quotes/{quote_id}/reject:
    post:
      tags:
      - quotes
      summary: Reject with an optional note
      operationId: reject_quote_api_v1_quotes__quote_id__reject_post
      security:
      - HTTPBearer: []
      parameters:
      - name: quote_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Quote Id
      - name: If-Match
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: If-Match
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuoteReject'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/QuoteRead'
                - $ref: '#/components/schemas/QuotePublicRead'
                title: Response Reject Quote Api V1 Quotes  Quote Id  Reject Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/quotes/{quote_id}/revise:
    post:
      tags:
      - quotes
      summary: New draft version copying the current content
      operationId: revise_quote_api_v1_quotes__quote_id__revise_post
      security:
      - HTTPBearer: []
      parameters:
      - name: quote_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Quote Id
      - name: If-Match
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: If-Match
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/QuoteRead'
                - $ref: '#/components/schemas/QuotePublicRead'
                title: Response Revise Quote Api V1 Quotes  Quote Id  Revise Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/quotes/{quote_id}/retry-email:
    post:
      tags:
      - quotes
      summary: Re-send the stored PDF after a failed delivery
      operationId: retry_email_api_v1_quotes__quote_id__retry_email_post
      security:
      - HTTPBearer: []
      parameters:
      - name: quote_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Quote Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/QuoteRead'
                - $ref: '#/components/schemas/QuotePublicRead'
                title: Response Retry Email Api V1 Quotes  Quote Id  Retry Email Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/opportunities/{opportunity_id}/quotes:
    get:
      tags:
      - quotes
      summary: Current quote versions of one opportunity
      operationId: list_opportunity_quotes_api_v1_opportunities__opportunity_id__quotes_get
      security:
      - HTTPBearer: []
      parameters:
      - name: opportunity_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Opportunity Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/QuoteSummaryRead'
                title: Response List Opportunity Quotes Api V1 Opportunities  Opportunity
                  Id  Quotes Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    OpportunityCreate:
      properties:
        account_id:
          type: string
          format: uuid
          title: Account Id
        division_id:
          type: string
          format: uuid
          title: Division Id
        estimated_amount:
          anyOf:
          - type: number
            minimum: 0.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*(?:\d{0,10}|(?=[\d.]{1,13}0*$)\d{0,10}\.\d{0,2}0*$)
          title: Estimated Amount
        pipeline_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Pipeline Id
        name:
          anyOf:
          - type: string
            maxLength: 200
            minLength: 1
          - type: 'null'
          title: Name
        description:
          anyOf:
          - type: string
            maxLength: 2000
          - type: 'null'
          title: Description
        expected_close_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Expected Close Date
        is_tender:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Tender
        tender_reference:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
          title: Tender Reference
        tender_deadline:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Tender Deadline
        estimated_award_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Estimated Award Date
        owner_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Owner Id
      type: object
      required:
      - account_id
      - division_id
      - estimated_amount
      title: OpportunityCreate
...
    Page_OpportunitySummaryRead_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/OpportunitySummaryRead'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        page_size:
          type: integer
          title: Page Size
      type: object
      required:
      - items
      - total
      - page
      - page_size
      title: Page[OpportunitySummaryRead]
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer